requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
streamlit-flow-component>=1.2.0
GitPython>=3.1.40